+++ source/android/adaptivecards/src/main/java/io/adaptivecards/renderer/action/ActionElementRenderer.java
@@ -29,7 +29,7 @@ import io.adaptivecards.objectmodel.HostConfig;
 import io.adaptivecards.objectmodel.IconPlacement;
 import io.adaptivecards.objectmodel.SubmitAction;
 import io.adaptivecards.renderer.BaseActionElementRenderer;
-import io.adaptivecards.renderer.FluentIconUtils;
+import io.adaptivecards.renderer.IconUtils;
 import io.adaptivecards.renderer.RenderArgs;
 import io.adaptivecards.renderer.RenderedAdaptiveCard;
 import io.adaptivecards.renderer.Util;
@@ -216,7 +216,7 @@ public class ActionElementRenderer extends BaseActionElementRenderer
             inputHandler.addInputWatcher((id, val) -> {
                 boolean isEnabled = isAnyInputValid(inputHandlers);
                 button.setEnabled(isEnabled);
-                FluentIconUtils.applyIconColor(button, button.getCurrentTextColor());
+                IconUtils.applyIconColor(button, button.getCurrentTextColor());
             });
         }
     }
